refactor: interface 파일로 분리
