+++ components/src/lib/Modal.tsx
@@ -1,25 +1,9 @@
-import { ReactNode, MouseEvent, CSSProperties, useEffect } from "react";
+import { MouseEvent, PropsWithChildren, useEffect } from "react";
 import { Modal } from "../lib";
 import "./Modal.css";
-import { Size, Type, Position } from "./type/modalType";
+import { ModalMainProps } from "./type/modal.type";
 
-interface ModalMainProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onConfirm?: () => void;
-  onCancel?: () => void;
-  onSubmit?: () => void;
-  type?: Type;
-  size?: Size;
-  position?: Position;
-  className?: string;
-  zIndex?: number;
-  customStyle?: CSSProperties;
-  children?: ReactNode;
-  portalRoot?: HTMLElement | null;
-}
-
-export const ModalMain = ({ isOpen, onClose, onConfirm, onCancel, onSubmit, type = "default", size = "medium", position = "center", className = "", zIndex = 999, customStyle = {}, children, portalRoot = document.body }: ModalMainProps) => {
+export const ModalMain = ({ isOpen, onClose, onConfirm, onCancel, onSubmit, type = "default", size = "medium", position = "center", className = "", zIndex = 999, children, portalRoot = document.body, ...rest }: PropsWithChildren<ModalMainProps>) => {
   useEffect(() => {
     if (portalRoot) {
       portalRoot.style.overflow = isOpen ? "hidden" : "auto";
@@ -33,14 +17,17 @@ export const ModalMain = ({ isOpen, onClose, onConfirm, onCancel, onSubmit, type
   if (!isOpen) return null;
 
   return (
-    <div className="modal-container">
+    <div
+      className="modal-container"
+      {...rest}
+    >
       <div
         className="back-drop"
         onClick={onClose}
       ></div>
       <div
         className={`modal-content-container ${position} ${size} ${className}`}
-        style={{ zIndex, ...customStyle }}
+        style={{ zIndex }}
         onClick={handleModalContainerClick}
       >
         {children}
