+++ components/src/lib/modalInput/ModalInput.tsx
@@ -1,27 +1,12 @@
-import { ReactNode, CSSProperties } from "react";
+import { ModalInputProps } from "../type/modal.type";
 import "./ModalInput.css";
 
-interface ModalInputProps {
-  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  type?: string;
-  value?: string;
-  placeholder?: string;
-  children?: ReactNode;
-  className?: string;
-  customStyle?: CSSProperties;
-}
-
-const ModalInput = ({ onChange, type = "string", placeholder = "", className = "", customStyle = {}, children }: ModalInputProps) => {
+const ModalInput = ({ ...rest }: ModalInputProps) => {
   return (
     <input
-      onChange={onChange}
-      type={type}
-      placeholder={placeholder}
-      className={`modal-input ${className}`}
-      style={customStyle}
-    >
-      {children}
-    </input>
+      className="modal-input"
+      {...rest}
+    />
   );
 };
