--- components/src/lib/type/modalType.ts
@@ -1,9 +1,0 @@
-export type Type = "alert" | "confirm" | "prompt" | "default";
-
-export type Size = "small" | "medium" | "large";
-
-export type Position = "center" | "bottom";
-
-export type ButtonVariant = "primary" | "secondary";
-
-export type ButtonSize = "small" | "large";
